compiler.h: add is_const() as a replacement of __is_constexpr()
__is_constexpr(), while being one of the most glorious one liner hack
ever witnessed by mankind, is overly complex. Following the adoption
of C11 in the kernel, this macro can be simplified through the use of
a _Generic() selection.

First, split the macro in two:

  - __is_const_zero(x): an helper macro; tells whether x is the
    integer constant expression 0 or something else.

  - is_const(x): replacement of __is_constexpr(); tells whether x is a
    integer constant expression.

The split serves two purposes: first make it easier to understand;
second, __is_const_zero() will be reused as a building block for other
is_const_*() macros that will be introduced later on.

The core principle of __is_constexpr() to abuse the return type of the
ternary operator remains, but all the surrounding sizeof() hack
disappear.

is_const() is designed to only accept integer arguments. The main
reason is that, by definition, a *pointer* can not be an *integer*
constant expression. Indeed, in the previous implementation,
__is_constexpr(NULL) returns false and, so far, no one ever used
__is_constexpr() with a pointer argument. Rather that allowing pointers
and returning a misleading result, ban this use case by creating a
constraint violation (multiplication between int and pointer types).

On a side note, while not relevant to the kernel, __is_constexpr()
relied on the GNU extension that sizeof(void) is 1. const_expr() does
not use any GNU extensions, making it ISO C compliant.

__is_constexpr() is temporarily kept and will be removed once all its
users get migrated to is_const() (or one of its friend).

Suggested-by: David Laight <David.Laight@aculab.com>
Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>
---
Changes in v2:

  - Remove the homage to Martin, instead, an entry in the CREDITS file
    will be added the last patch of this series.

  - Rephrase the documentation and the commit comments. Explain why
    pointer arguments are banned.