compiler.h: add is_const_true()
statically_true() may not be usable in all contexts. For example:

  void foo(int a)
  {
  	 /* fail on GCC */
  	BUILD_BUG_ON_ZERO(statically_true(a));

  	 /* fail on both clang and GCC */
  	static char arr[statically_true(a) ? 1 : 2];
  }

Define the new is_const_true() macro which, by making use of
__is_const_zero(), always produces a constant expression.

Note that using the ! operator like this:

  #define is_const_true(x) __is_const_zero(!(x))

would yield a -Wint-in-bool-context compiler warning if the argument
is not a boolean. Use the == operator instead.

It should be noted that statically_true() is the only one capable of
folding tautological expressions in which at least one of the operands
is not an integer constant expression. For example:

  statically_true(true || var)
  statically_true(var == var)
  statically_true(var * 0 == 0)
  statically_true(var * 8 % 4 == 0)

always evaluate to true, whereas all of these would be false under
is_const_true() if var is not a constant expression [3].

For this reason, usage of const_true() should be the exception.
Reflect const_true()'s limitation in the documentation and state that
statically_true() is the overall preferred solution.

[1] __builtin_constant_p cannot resolve to const when optimizing
Link: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=19449

[2] commit 3c8ba0d61d04 ("kernel.h: Retain constant expression output for max()/min()")
Link: https://git.kernel.org/torvalds/c/3c8ba0d61d04

[3] https://godbolt.org/z/E3TPnP3qf

Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>
---
Changes in v2:

  - remove is_const_false().

  - remove the introduction paragraph in the commit comment.

  - rephrase the documentation block to drop the term "similar".

  - macro definitions in godbolt link (footnote [3]) used an earlier
    version and could yield warnings. Update the macros in the godbolt
    link to match those in the patch. The old godbold link was:

      https://godbolt.org/z/E4r7EaxW9