drm/i915/sprite: fix brightness value
The brightness is set to a negative number. When passed to
REG_FIELD_PREP(), the value will is interpreted as an unsigned
(c.f. __builtin_ffsll(x)), resulting in a value too large for the
field.

SP_BRIGHTNESS() is documented as being an s8, so the intent seems
indeed for this value to be negative. Declare brightness as an u8 so
that it matches the destination width and so that no bits are lost when
preparing the bitfield.

This has no functional change but will prevent a build bug error when
REG_FIELD_PREP() will be refactored to perform better static analysis
in an upcoming change.

Fixes: 5deae9191130 ("drm/i915: Correctly handle limited range YCbCr data on VLV/CHV")
Signed-off-by: Vincent Mailhol <mailhol.vincent@wanadoo.fr>
---
FYI, my tests: https://godbolt.org/z/frYMK4v8M