can: netlink: add CAN XL
Because of popular request [1] and to celebrate the fact that I joined
the kernel web of trust this week [2], here is an RFC to introduce the
CAN XL netlink support.

The logic is simple. The CAN XL basically just reuse the same kind of
parameters as CAN FD: the databittiming and the TDC. So the series is
just a bunch of refactor to either:

  - factorize code.
  - rename some variable to differentiate between CAN FD and XL.
  - make some function more generic to accept both CAN FD and XL.

The feature is working™: there is a dummy driver at the end of the
series to show the traffic from the userland to a driver. This said, I
did close to zero testing. Once I had one CAN XL frame reaching the
driver, I call it a day, and decided to send the work. Regardless, it
is Sunday night. If I do not send it now, that would be next week-end.
Probably some mistakes are hidden here and there, but this should be
enough for an RFC level.

Also, I am not fully happy that can_dbt_changelink() requires 8
parameters. I will probably revisit this later on. But for the moment,
I think this is acceptable for an RFC.

Overall, I do not want to rush this series. Linus should send the rc7
anytime soon, and the merge window will probably start in eight days.
I do not think this series will be finalized by then. I still need to
give a deeper look at ISO 11898-1:2024 [3] to check that everything is
good. However, if I receive positive feedback on this RFC, I would
probably like to have the first patch merged so that I do not have to
rebase that tree wide patch each time someone makes a change.

I will send a second RFC series for iplink2 just after this one. Stay
tuned!

[1] https://lore.kernel.org/linux-can/CAEQ16vpxthctdrpv0kBKEZJA8VNYffjGGPBGBY93RmKDD49bAQ@mail.gmail.com/
[2] https://git.kernel.org/pub/scm/docs/kernel/pgpkeys.git/commit/?id=81d335d3586c
[3] https://www.iso.org/fr/standard/86384.html

To: Marc Kleine-Budde <mkl@pengutronix.de>
To: Oliver Hartkopp <socketcan@hartkopp.net>
To: linux-can@vger.kernel.org
Cc: Robert Nawrath <mbro1689@gmail.com>
Cc: Pavel Pisa <pisa@cmp.felk.cvut.cz>
Cc: Ondrej Ille <ondrej.ille@gmail.com>
Cc: Chandrasekar Ramakrishnan <rcsekar@samsung.com>
Cc: kernel@pengutronix.de
Cc: Heiko Stuebner <heiko@sntech.de>
Cc: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>
Cc: Thomas Kopp <thomas.kopp@microchip.com>
Cc: Frank Jungclaus <frank.jungclaus@esd.eu>
Cc: socketcan@esd.eu
Cc: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>
Cc: Michal Simek <michal.simek@amd.com>
Cc: linux-kernel@vger.kernel.org
Cc: linux-arm-kernel@lists.infradead.org
Cc: linux-rockchip@lists.infradead.org
Signed-off-by: Vincent Mailhol <mailhol@kernel.org>

--- b4-submit-tracking ---
{
  "series": {
    "revision": 2,
    "change-id": "20241229-canxl-netlink-bc640af10673",
    "prefixes": [
      "RFC"
    ]
  }
}