can: netlink: add can_dtb_changelink()
Factorize the dattabittiming parsing out of can_changelink() and move
it in the new can_dtb_changelink() function. This is a preparation
patch for the introduction of CAN XL because the databittiming
changelink logic will be reused later on.

Signed-off-by: Vincent Mailhol <mailhol@kernel.org>
---
I am not fully happy that can_dbt_changelink() requires 8 parameters.
I will probably revisit this later on. But for the moment, I think
this is acceptable for an RFC.