minmax: refactor __is_nonneg() using GNU's compound statement expression
The __is_nonneg() function like macros in mimmax can use GNU's
compound statement expressions. This allows to rewrite them in a more
intelligible way.

First, replace the:

  #if __SIZEOF_POINTER__ == __SIZEOF_LONG_LONG__

preprocessor check by a check on the intptr_t type:

  sizeof(ux) <= sizeof(intptr_t)

This replacement allows to detect in a portable way when a pointer
cast is needed, thus removing the need for two different
implementations.

Second, replace the documentation paragraph which states that signed
128 types are not supported by a BUILD_BUG_ON_MSG() check so that
misuses are caught at compile time. Even though s128 are not used as
of today, someone might try to use the minmax.h macros with that type
one day.

Third and final, the current documentation claim that the cast will
remove any warning. This is not entirely true: at W=2, GCC still emit
a -Wtype-limits warning diagnostic when an unsigned value is casted to
a wider signed type, for example:

  void foo(unsigned u)
  {
	(long long)u >= 0
  }

Assigning the casted value to a temporary variable resolves this
warning.

Signed-off-by: Vincent Mailhol <mailhol@kernel.org>